Separated Messenger feature.
