+++ app/src/main/java/com/nailorsh/repeton/data/repositories/FakeRepetonRepository.kt
@@ -1,9 +1,7 @@
 package com.nailorsh.repeton.data.repositories
 
-import com.nailorsh.repeton.data.sources.FakeChatsSource
 import com.nailorsh.repeton.data.sources.FakeLessonSource
 import com.nailorsh.repeton.domain.repositories.RepetonRepository
-import com.nailorsh.repeton.model.Chat
 import com.nailorsh.repeton.model.Lesson
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
@@ -11,9 +9,7 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class FakeRepetonRepository @Inject constructor() : RepetonRepository {
-    override suspend fun getChats(): List<Chat> = withContext(Dispatchers.IO) {
-        FakeChatsSource.generateChatList()
-    }
+
 
     override suspend fun getLesson(id: Int): Lesson = withContext(Dispatchers.IO) {
         delay(2000)
