+++ app/src/main/java/com/nailorsh/repeton/domain/repositories/MessengerRepository.kt
@@ -1,0 +1,7 @@
+package com.nailorsh.repeton.domain.repositories
+
+import com.nailorsh.repeton.model.Chat
+
+interface MessengerRepository {
+    suspend fun getChats(): List<Chat>
+}
